placed classes in their proper places
