+++ src/exceptions/UserException.java
@@ -1,4 +1,4 @@
-package Exceptions;
+package exceptions;
 
 public class UserException extends Exception {
     public UserException(String message) {
