+++ src/model/playerclasses/Mage.java
@@ -1,13 +1,18 @@
-package model.playerClasses;
+package model.playerclasses;
 
 import enums.PlayerClass;
 
-public class Assassin implements SetStats {
+public class Mage implements SetStats {
     private int attack = 20;
-    private int defense = 30;
-    private int health = 70;
-    private int specialAttack = 110;
-    private int specialDefense = 70;
+    private int defense = 50;
+    private int health = 60;
+    private int specialAttack = 120;
+    private int specialDefense = 50;
+
+    @Override
+    public void SetPlayerStatsForClass(PlayerClass playerClass) {
+        System.out.println("The players class : " + playerClass);
+    }
 
     public int getAttack() {
         return attack;
@@ -28,10 +33,4 @@ public class Assassin implements SetStats {
     public int getSpecialDefense() {
         return specialDefense;
     }
-
-    @Override
-    public void SetPlayerStatsForClass(PlayerClass playerClass) {
-        System.out.println("Player Class: " + playerClass);
-    }
 }
-
